pyproject.toml
[project]
name = "ai-based-form"
version = "0.1.0"
description = "AI-powered friendly questionnaire with Google Sheets integration"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "google-genai>=1.0.0",
    "gspread>=6.1.0",
    "google-auth>=2.35.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "pydantic>=2.9.0",
]
